+++ runtime/include/tt/runtime/detail/distributed/controller/controller.h
@@ -114,7 +114,7 @@ public:
 
   ::tt::runtime::Tensor createOwnedHostTensor(
       const void *data, const std::vector<std::uint32_t> &shape,
-      const std::vector<std::uint32_t> &stride, std::uint32_t itemsize,
+      const std::vector<std::uint64_t> &stride, std::uint32_t itemsize,
       ::tt::target::DataType dataType);
 
   ::tt::runtime::Tensor createMultiDeviceHostTensor(
